fix: 读取buff时，长度被ushort截断的bug

+++ Assets/SSQA/Kits/RemoteDebugger/Client/NetClient.cs
@@ -69,8 +69,8 @@ public class NetClient : IDisposable {
                         
                         byRead.CopyTo(by_data, offset);
 
-                        offset += (ushort)read_byte_count;
-                        data_len -= (ushort)read_byte_count;
+                        offset += read_byte_count;
+                        data_len -= read_byte_count;
                     }
                     catch (Exception ex) {
                         Debug.LogException(ex);
